Rename `Api.V2` to `Api`.
