+++ src/Cardano/Wallet/Api.hs
@@ -1,15 +1,15 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Cardano.Wallet.Api.V2 where
+module Cardano.Wallet.Api where
 
 import Data.Proxy
     ( Proxy (..) )
 import Servant.API
     ( (:<|>), (:>), Capture, Delete, Get, JSON )
 
-import qualified Cardano.Wallet.Api.V2.Types.Wallet as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletId as T
+import qualified Cardano.Wallet.Api.Types.Wallet as T
+import qualified Cardano.Wallet.Api.Types.WalletId as T
 
 type Api = DeleteWallet :<|> GetWallet :<|> ListWallets
