+++ src/Cardano/Wallet/Api/Types/Wallet.hs
@@ -1,23 +1,23 @@
 {-# LANGUAGE DeriveGeneric #-}
 
-module Cardano.Wallet.Api.V2.Types.Wallet where
+module Cardano.Wallet.Api.Types.Wallet where
 
 import Prelude
 
-import Cardano.Wallet.Api.V2.JSON
+import Cardano.Wallet.Api.JSON
     ( defaultRecordTypeOptions )
 import Data.Aeson
     ( FromJSON (..), ToJSON (..), genericParseJSON, genericToJSON )
 import GHC.Generics
     ( Generic )
 
-import qualified Cardano.Wallet.Api.V2.Types.WalletAddressPoolGap as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletBalance as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletDelegation as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletId as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletName as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletPassphraseInfo as T
-import qualified Cardano.Wallet.Api.V2.Types.WalletState as T
+import qualified Cardano.Wallet.Api.Types.WalletAddressPoolGap as T
+import qualified Cardano.Wallet.Api.Types.WalletBalance as T
+import qualified Cardano.Wallet.Api.Types.WalletDelegation as T
+import qualified Cardano.Wallet.Api.Types.WalletId as T
+import qualified Cardano.Wallet.Api.Types.WalletName as T
+import qualified Cardano.Wallet.Api.Types.WalletPassphraseInfo as T
+import qualified Cardano.Wallet.Api.Types.WalletState as T
 
 data Wallet = Wallet
     { _id :: !T.WalletId
