+++ src/Cardano/Wallet/Api/Types/WalletBalance.hs
@@ -1,17 +1,17 @@
 {-# LANGUAGE DeriveGeneric #-}
 
-module Cardano.Wallet.Api.V2.Types.WalletBalance where
+module Cardano.Wallet.Api.Types.WalletBalance where
 
 import Prelude
 
-import Cardano.Wallet.Api.V2.JSON
+import Cardano.Wallet.Api.JSON
     ( defaultRecordTypeOptions )
 import Data.Aeson
     ( FromJSON (..), ToJSON (..), genericParseJSON, genericToJSON )
 import GHC.Generics
     ( Generic )
 
-import qualified Cardano.Wallet.Api.V2.Types.Amount as T
+import qualified Cardano.Wallet.Api.Types.Amount as T
 
 data WalletBalance = WalletBalance
     { _available :: !T.Amount
